Add: support update item
